requirements.txt
# Textil PRO Bot Dependencies
pyTelegramBotAPI==4.14.0
zep-cloud==2.14.1
aiohttp==3.9.1
aiofiles==24.1.0
python-dotenv==1.0.0
openai==1.3.7
httpx==0.24.1

# Webhook and Business API Support
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2

# Streamlit Admin Panel
streamlit==1.28.1
streamlit-ace==0.1.1
gitpython==3.1.40
requests==2.31.0